Conflicts:
	src/core/IIndexBuffer.h
	src/core/IVertexBuffer.h
	src/render/OGL1_3/CIndexBufferOGL1_3.h
	src/render/OGL1_3/CVertexBufferOGL1_3.h
	src/render/OGL1_5/CIndexBufferOGL1_5.h
	src/render/OGL1_5/CVertexBufferOGL1_5.h
	src/render/OGL2_0/CIndexBufferOGL2_0.h
	src/render/OGL2_0/CVertexBufferOGL2_0.h

+++ src/ces/IEntity.cpp
@@ -111,7 +111,7 @@ bool IEntity::HasComponent(const std::type_index& id)
     return !components.empty();
 }
 
-size_t IEntity::ComponentsNum(const std::type_index& id)
+uint32_t IEntity::ComponentsNum(const std::type_index& id)
 {
     TComponentsList& components = m_Components[id];
     return components.size();
